Cleanup mutexes
